fix: relocate fusion-watchdog out of plugins dir to fix standalone plugin crash
Amp scans every .ts file in ~/.config/amp/plugins as a standalone plugin.
The watchdog helper module has no default export, causing:
  Plugin must export a default function
  Plugin process exited before readiness

Move fusion-watchdog.ts from configs/amp/plugins/ to configs/amp/lib/.
Update the installer (cli.sh) and exporter (generate.sh) to copy the
lib/ directory alongside plugins. The relative import in fusion-agents.ts
changes from ./fusion-watchdog to ../lib/fusion-watchdog, which resolves
correctly at the installed path ~/.config/amp/plugins/../lib/.

Add regression tests:
- Every .ts in configs/amp/plugins/ must export a default function
- fusion-watchdog.ts must NOT exist in configs/amp/plugins/
- Installer must copy configs/amp/lib/ to ~/.config/amp/lib/

Amp-Thread-ID: https://ampcode.com/threads/T-019fc5b3-6f92-7742-987d-7a07d9fc169b
