+++ configs/amp/lib/fusion-watchdog.ts
